chore(wallaby): fix broken tests and consolidate configs
