--- wallaby.form-foundations.js
@@ -1,6 +1,0 @@
-// TODO: move to wallaby.config.js
-const wallabyBase = require('./wallaby.base');
-module.exports = wallabyBase(
-  ['packages/*/src/**'],
-  ['packages/*/src/**/*.spec.{ts,tsx}'],
-);
